pyfastani/tests/requirements.txt
biopython
scikit-bio
